requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
pgvector>=0.2.5

httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
readability-lxml>=0.8.1
markdownify>=0.13.1  # HTML to Markdown conversion with table support

tenacity>=8.2.3

langchain==1.1.0
langchain-openai>=0.1.0
langfuse>=2.38.0

# cross-encoder reranker（bge-reranker）
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
